+++ lib/tar.mli
@@ -19,9 +19,10 @@
     {e %%VERSION%% - {{:%%PKG_HOMEPAGE%% }homepage}} *)
 
 module Header : sig
-  (** Process and create tar file headers *)
+  (** Process and create tar file headers. *)
 
-  (** tar format assumptions. Default is [V7] (for compatibility with versions of ocaml-tar before this type was introduced). See http://www.gnu.org/software/tar/manual/html_section/tar_68.html for more information. *)
+  (** tar format assumptions. Default is {!V7} (for compatibility with versions of ocaml-tar before this type was introduced).
+      @see <http://www.gnu.org/software/tar/manual/html_section/tar_68.html> *)
   type compatibility =
     | OldGNU (** GNU tar < 1.12 *)
     | GNU (** GNU tar 1.12 - 1.13.25 *)
@@ -29,11 +30,11 @@ module Header : sig
     | Ustar (** POSIX.1-1988 *)
     | Posix (** POSIX.1-2001 *)
 
-  (** Default compatibility if [?level] is omitted. Defaults to [V7] *)
+  (** Default compatibility if [?level] is omitted. Defaults to {!V7}. *)
   val compatibility_level : compatibility ref
 
   module Link : sig
-    (** Determines the type of the file *)
+    (** Determines the type of the file. *)
     type t =
       | Normal
       | Hard (** a hard link *)
@@ -63,12 +64,12 @@ module Header : sig
       user_id: int option;
       uname: string option;
     }
-    (** Represents a "Pax" extended header *)
+    (** Represents a "Pax" extended header. *)
 
     val unmarshal : Cstruct.t -> t
   end
 
-  (** Represents a standard archive (note checksum not stored) *)
+  (** Represents a standard archive (note checksum not stored). *)
   type t = {
     file_name : string;
     file_mode: int;
@@ -85,43 +86,43 @@ module Header : sig
     extended: Extended.t option;
   }
 
-  (** Helper function to make a simple header *)
+  (** Helper function to make a simple header. *)
   val make : ?file_mode:int -> ?user_id:int -> ?group_id:int -> ?mod_time:int64 -> ?link_indicator:Link.t -> ?link_name:string -> ?uname:string -> ?gname:string -> ?devmajor:int -> ?devminor:int -> string -> int64 -> t
 
-  (** Length of a header block *)
+  (** Length of a header block. *)
   val length : int
 
-  (** A blank header block (two of these in series mark the end of the tar) *)
+  (** A blank header block (two of these in series mark the end of the tar). *)
   val zero_block : Cstruct.t
 
-  (** Pretty-print the header record *)
+  (** Pretty-print the header record. *)
   val to_detailed_string : t -> string
 
-  (** For debugging: pretty-print a string as hex *)
+  (** For debugging: pretty-print a string as hex. *)
   val to_hex : string -> string
 
-  (** Thrown when unmarshalling a header if the checksums don't match *)
+  (** Thrown when unmarshalling a header if the checksums don't match. *)
   exception Checksum_mismatch
 
-  (** Thrown if we detect the end of the tar (at least two zero blocks in sequence) *)
+  (** Thrown if we detect the end of the tar (at least two zero blocks in sequence). *)
   exception End_of_stream
 
-  (** Unmarshal a header block, returning None if it's all zeroes.
-      This header block may be preceeded by an [?extended] block which
+  (** Unmarshal a header block, returning [None] if it's all zeroes.
+      This header block may be preceded by an [?extended] block which
       will override some fields. *)
   val unmarshal : ?level:compatibility -> ?extended:Extended.t -> Cstruct.t -> t option
 
-  (** Marshal a header block, computing and inserting the checksum *)
+  (** Marshal a header block, computing and inserting the checksum. *)
   val marshal : ?level:compatibility -> Cstruct.t -> t -> unit
 
   (** Compute the amount of zero-padding required to round up the file size
-      to a whole number of blocks *)
+      to a whole number of blocks. *)
   val compute_zero_padding_length : t -> int
 
-  (** Return the required zero-padding as a string *)
+  (** Return the required zero-padding as a string. *)
   val zero_padding : t -> Cstruct.t
 
-  (** [to_sectors t] is the number of sectors occupied by the data *)
+  (** [to_sectors t] is the number of sectors occupied by the data. *)
   val to_sectors: t -> int64
 end
 
@@ -146,10 +147,10 @@ end
 
 module HeaderReader(Async: ASYNC)(Reader: READER with type 'a t = 'a Async.t) :
  sig
-  (** Returns the next header block or throws End_of_stream if two consecutive
+  (** Returns the next header block or throws {!Header.End_of_stream} if two consecutive
       zero-filled blocks are discovered. Assumes stream is positioned at the
-      possible start of a header block. End_of_file is thrown if the stream
-      unexpectedly fails *)
+      possible start of a header block.
+      @raise Header.End_of_stream if the stream unexpectedly fails. *)
   val read : ?level:Header.compatibility -> Reader.in_channel -> (Header.t, [`Eof]) result Async.t
 end
 
@@ -170,19 +171,19 @@ end
 module Make (IO : IO) : sig
   val really_read: IO.in_channel -> Cstruct.t -> unit
   (** [really_read fd buf] fills [buf] with data from [fd] or raises
-      End_of_file *)
+      {!Stdlib.End_of_file}. *)
 
   val really_write: IO.out_channel -> Cstruct.t -> unit
   (** [really_write fd buf] writes the full contents of [buf] to [fd]
-      or raises End_of_file *)
+      or raises {!Stdlib.End_of_file}. *)
 
   module Header : sig
     include module type of Header
 
-    (** Returns the next header block or fails with `Eof if two consecutive
+    (** Returns the next header block or fails with [`Eof] if two consecutive
         zero-filled blocks are discovered. Assumes stream is positioned at the
-        possible start of a header block. End_of_file is thrown if the stream
-        unexpectedly fails *)
+        possible start of a header block.
+        @raise Stdlib.End_of_file if the stream unexpectedly fails. *)
     val get_next_header : ?level:compatibility -> IO.in_channel -> t
   end
 
@@ -193,31 +194,31 @@ module Make (IO : IO) : sig
 
   val write_end: IO.out_channel -> unit
     [@@ocaml.deprecated "Deprecated: use Tar.HeaderWriter"]
-  (** Writes a stream terminator to [fd]  *)
+  (** Writes a stream terminator to [fd]. *)
 
   module Archive : sig
-    (** Utility functions for operating over whole tar archives *)
+    (** Utility functions for operating over whole tar archives. *)
 
     (** Read the next header, apply the function 'f' to the fd and the header. The function
         should leave the fd positioned immediately after the datablock. Finally the function
-        skips past the zero padding to the next header *)
+        skips past the zero padding to the next header. *)
     val with_next_file : IO.in_channel -> (IO.in_channel -> Header.t -> 'a) -> 'a
 
-    (** List the contents of a tar *)
+    (** List the contents of a tar. *)
     val list : ?level:Header.compatibility -> IO.in_channel -> Header.t list
 
     (** [extract_gen dest] extract the contents of a tar.
-        Apply 'dest' on each header to get a handle to the file to write to *)
+        Apply [dest] on each header to get a handle to the file to write to. *)
     val extract_gen : (Header.t -> IO.out_channel) -> IO.in_channel -> unit
 
-    (** Create a tar on file descriptor fd from the stream of headers.  *)
+    (** Create a tar on file descriptor fd from the stream of headers. *)
     val create_gen : ?level:Header.compatibility -> (Header.t * (IO.out_channel -> unit)) Stream.t -> IO.out_channel -> unit
 
-    (** [copy_n ifd odf n] copies exactly [n] bytes from [ifd] to [ofd] *)
+    (** [copy_n ifd odf n] copies exactly [n] bytes from [ifd] to [ofd]. *)
     val copy_n : IO.in_channel -> IO.out_channel -> int64 -> unit
       [@@ocaml.deprecated "Deprecated: use your own helper function"]
 
-    (** [skip fd n] reads and throws away [n] bytes from [fd] *)
+    (** [skip fd n] reads and throws away [n] bytes from [fd]. *)
     val skip : IO.in_channel -> int -> unit
       [@@ocaml.deprecated "Deprecated: use your own helper function"]
   end