+++ CHANGES.md
@@ -1,3 +1,9 @@
+## unreleased
+
+- Bump lower-bound on Cstruct to 6.0.0 (@MisterDA)
+- Update to Dune 2.9 and generate opam files (@MisterDA)
+- Don't print the name of the file in extract functions (@MisterDA)
+
 ## v1.1.0 (2019-04-08)
 
 - Do not depend on mirage-types, use mirage-kv instead (@hannesm)